Added options to the table creatioon script
o - label, label to display for the field
o - description, end user description of what the field is for.
o - cacheindex, whether to cache the field for queries, otherwise read the file directly, we will not cache textarea/code etc.. fields as they may use too much memory
o - tableview, whether to show the field on the record management overview.